Create link view in text VH using attachment view factory
