+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/MessageListItemViewHolderFactory.kt
@@ -96,7 +96,8 @@ public open class MessageListItemViewHolderFactory(
             parentView,
             decoratorProvider.decorators,
             listenerContainer,
-            markdown
+            markdown,
+            attachmentViewFactory
         )
     }
