+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/internal/MessageListItemDecoratorProvider.kt
@@ -9,7 +9,6 @@ import io.getstream.chat.android.ui.message.list.adapter.viewholder.decorator.in
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.decorator.internal.FailedIndicatorDecorator
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.decorator.internal.FootnoteDecorator
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.decorator.internal.GapDecorator
-import io.getstream.chat.android.ui.message.list.adapter.viewholder.decorator.internal.LinkAttachmentDecorator
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.decorator.internal.MaxPossibleWidthDecorator
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.decorator.internal.ReactionsDecorator
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.decorator.internal.ReplyDecorator
@@ -31,7 +30,6 @@ internal class MessageListItemDecoratorProvider(
         AvatarDecorator(),
         FailedIndicatorDecorator(),
         ReactionsDecorator(),
-        LinkAttachmentDecorator(),
         ReplyDecorator(currentUser),
         ReactionsDecorator(),
         FootnoteDecorator(dateFormatter, isDirectMessage),
