+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/viewholder/internal/PlainTextWithFileAttachmentsViewHolder.kt
@@ -2,8 +2,8 @@ package io.getstream.chat.android.ui.message.list.adapter.viewholder.internal
 
 import android.view.ViewGroup
 import androidx.core.view.isVisible
-import com.getstream.sdk.chat.ChatMarkdown
 import androidx.core.view.setPadding
+import com.getstream.sdk.chat.ChatMarkdown
 import com.getstream.sdk.chat.adapter.MessageListItem
 import com.getstream.sdk.chat.utils.extensions.inflater
 import io.getstream.chat.android.client.extensions.uploadId
