+++ stream-chat-android-ui-components/src/main/res/layout/stream_ui_item_message_plain_text.xml
@@ -62,13 +62,10 @@
             android:visibility="gone"
             />
 
-        <io.getstream.chat.android.ui.message.list.adapter.view.internal.LinkAttachmentView
-            android:id="@+id/linkAttachmentView"
+        <FrameLayout
+            android:id="@+id/attachmentsContainer"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:padding="@dimen/stream_ui_spacing_small"
-            android:visibility="gone"
-            tools:visibility="visible"
             />
 
         <TextView
@@ -102,11 +99,12 @@
         android:id="@+id/threadGuideline"
         android:layout_width="0dp"
         android:layout_height="0dp"
-        android:orientation="horizontal"
         android:layout_marginBottom="@dimen/stream_ui_message_thread_reply_footnote_translation"
-        app:layout_constraintStart_toStartOf="parent"
+        android:orientation="horizontal"
+        app:layout_constraintBottom_toBottomOf="@id/messageContainer"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toBottomOf="@id/messageContainer"/>
+        app:layout_constraintStart_toStartOf="parent"
+        />
 
     <io.getstream.chat.android.ui.message.list.adapter.view.internal.FootnoteView
         android:id="@+id/footnote"
